requirements.txt
discord.py
pillow
pytesseract
requests
